update all auth strats
